+++ src/main/java/com/deviget/minesweeper/domain/model/game/cell/Cell.java
@@ -1,7 +1,5 @@
 package com.deviget.minesweeper.domain.model.game.cell;
 
-import org.springframework.data.annotation.PersistenceConstructor;
-
 import java.util.Objects;
 
 public abstract class Cell implements Comparable<Cell> {
@@ -22,7 +20,6 @@ public abstract class Cell implements Comparable<Cell> {
     return EmptyCell.at(index);
   }
 
-  @PersistenceConstructor
   protected Cell(final int index, final CellType type, final CellStatus status) {
     this.index = index;
     this.type = type;